Re-structured the project and database
Deleted the condition group table and class as there was no need to keep
it; the conditions are now related to the policy itself.
Deleted the tables and classes regarding the previous system that was
made to generate queries to be used for permissions as it was replaced
by the ABAC system.
Deleted the permissions table in the db and the corresponding class.
Re-factored the Setup service and DAO classes; the DAOs were deleted and
their methods were moved to the existing ABACDao while the service
classes were deleted too and their methods were transfered to the
SystemAbacService which is for in module use only and is being exported
to the testing module only.
The file module-info.java was added to the testing module improving the
whole module